Implement storedValueE which supports 'NeverRecomp'.
This commit introduces a new method to 'Rule', 'storedValueE',
whose return type is extended to be 'MaybeStored', a data type
that can reports 'AlwaysRecomp' (previously 'Nothing'),
'RecompIfNotEq' (previously 'Just') and 'NeverRecomp' (new!)
'NeverRecomp' causes Shake to treat the value stored for the
rule in the database as canonical; the rule will only be compiled
if (1) there was no value in the database to begin with, or
(2) dependencies changed.

See discussion in #191 (and #185).  This architectural extension
makes it possible to write oracle rules which do NOT always
rerun, although I haven't added any porcelain for that.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>